hyperlocal-backend/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.27,<1
sqlalchemy>=2.0,<3
pydantic>=2.6,<3
pydantic-settings>=2.2,<3
python-jose[cryptography]>=3.3,<4
bcrypt>=4.1,<5
python-multipart>=0.0.9,<1
slowapi>=0.1.9,<1
email-validator>=2.1,<3
httpx>=0.27,<1

pytest>=8,<9
